Fix glitches around MCP Servers flow
